+++ src/main/java/com/comanda/model/dto/EstoqueDto.java
@@ -1,9 +1,11 @@
 package com.comanda.model.dto;
 
+import java.math.BigDecimal;
+
 import lombok.Data;
 
 @Data
 public class EstoqueDto {
    private Long id;
-   private Integer quantidade;
+   private BigDecimal quantidade;
 }
